requirements.txt
streamlit
python-dotenv
requests
openai
arcadepy
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-openai
pandas
plotly
textual>=0.40.0
rich>=13.0.0
